removed config for rate limiter
