+++ backend/src/routes/v3/routes.ts
@@ -1,8 +1,6 @@
 import { Router } from 'express'
-import { rateLimit } from 'express-rate-limit'
 
 import { generateV3SwaggerSpec } from '../../services/specification.js'
-import config from '../../utils/config.js'
 import { getImageByDigest } from '../v3/model/images/getImage.js'
 import { getCurrentUser } from './entities/getCurrentUser.js'
 import { getComplianceMetrics } from './metrics/getComplianceMetrics.js'
@@ -12,18 +10,6 @@ import { getLatestResponse } from './response/getLatestResponseForReview.js'
 import { postNotifyReviewer } from './review/postNotifyReviewer.js'
 import { postReview } from './review/postReview.js'
 import { postReviewResponse } from './review/postReviewResponse.js'
-
-const reviewNotifierLimiter = rateLimit({
-  windowMs: config.smtp.review.lastNotifiedCoolDownMs,
-  limit: 1,
-  standardHeaders: true,
-  legacyHeaders: false,
-  keyGenerator: (req) => {
-    // Create a unique key based on the user and review ID so users can request different releases/access requests
-    const userDn = req.user?.dn ?? req.ip
-    return `${userDn}-${req.params['reviewId']}`
-  },
-})
 
 const router = Router()
 
@@ -38,7 +24,7 @@ router.get('/metrics/entryVolume', ...getEntryVolume)
 router.post('/review/:reviewId/response', ...postReviewResponse)
 router.post('/review/:modelId', ...postReview)
 router.get('/review/:reviewId/responses/latest', ...getLatestResponse)
-router.post('/review/:reviewId/notify', reviewNotifierLimiter, ...postNotifyReviewer)
+router.post('/review/:reviewId/notify', ...postNotifyReviewer)
 
 router.get('/entities/me', ...getCurrentUser)
