+++ backend/src/utils/config.ts
@@ -93,10 +93,6 @@ export interface Config {
       postReminderInterval: string
     }
 
-    review: {
-      lastNotifiedCoolDownMs: number
-    }
-
     from: string
   }
