+++ infrastructure/helm/bailo/values.yaml
@@ -289,7 +289,6 @@ config:
     transporter: "smtp"
     preReminderIntervals: ['1 day', '2 weeks', '10 weeks']
     postReminderInterval: '1 day'
-    lastNotifiedCoolDownMs: 300000
 
   app:
     protocol: "https"
